Refactor query and msg servers (#70)
