+++ x/emissions/keeper/msgserver/msg_server.go
@@ -1,0 +1,17 @@
+package msgserver
+
+import (
+	state "github.com/allora-network/allora-chain/x/emissions"
+	"github.com/allora-network/allora-chain/x/emissions/keeper"
+)
+
+type msgServer struct {
+	k keeper.Keeper
+}
+
+var _ state.MsgServer = msgServer{}
+
+// NewMsgServerImpl returns an implementation of the module MsgServer interface.
+func NewMsgServerImpl(keeper keeper.Keeper) state.MsgServer {
+	return &msgServer{k: keeper}
+}
