+++ x/emissions/keeper/msgserver/msg_server_forecasts_test.go
@@ -1,0 +1,1 @@
+package msgserver_test
