+++ x/emissions/keeper/msgserver/msg_server_inferences.go
@@ -1,0 +1,38 @@
+package msgserver
+
+import (
+	"context"
+
+	state "github.com/allora-network/allora-chain/x/emissions"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+// A tx function that accepts a list of inferences and possibly returns an error
+func (ms msgServer) ProcessInferences(ctx context.Context, msg *state.MsgProcessInferences) (*state.MsgProcessInferencesResponse, error) {
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+
+	inferences := msg.Inferences
+	// Group inferences by topicId - Create a map to store the grouped inferences
+	groupedInferences := make(map[uint64][]*state.Inference)
+
+	// Iterate through the array and group by topic_id
+	for _, inference := range inferences {
+		groupedInferences[inference.TopicId] = append(groupedInferences[inference.TopicId], inference)
+	}
+
+	actualTimestamp := uint64(sdkCtx.BlockTime().Unix())
+
+	// Update all_inferences
+	for topicId, inferences := range groupedInferences {
+		inferences := &state.Inferences{
+			Inferences: inferences,
+		}
+		err := ms.k.InsertInferences(ctx, topicId, actualTimestamp, *inferences)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	// Return an empty response as the operation was successful
+	return &state.MsgProcessInferencesResponse{}, nil
+}
