+++ x/emissions/keeper/msgserver/msg_server_losses_test.go
@@ -1,0 +1,57 @@
+package msgserver_test
+
+import (
+	cosmosMath "cosmossdk.io/math"
+	state "github.com/allora-network/allora-chain/x/emissions"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+func (s *KeeperTestSuite) TestMsgSetLosses() {
+	ctx, msgServer := s.ctx, s.msgServer
+	require := s.Require()
+
+	// Mock setup for addresses
+	reputerAddr := sdk.AccAddress(PKS[0].Address()).String()
+	workerAddr := sdk.AccAddress(PKS[1].Address()).String()
+
+	// Create a MsgSetLosses message
+	weightMsg := &state.MsgSetLosses{
+		Sender: reputerAddr,
+		LossBundles: []*state.LossBundle{
+			{
+				TopicId: 1,
+				Reputer: reputerAddr,
+				Worker:  workerAddr,
+				Weight:  cosmosMath.NewUint(100),
+			},
+		},
+	}
+
+	_, err := msgServer.SetWeights(ctx, weightMsg)
+	require.NoError(err, "SetWeights should not return an error")
+}
+
+func (s *KeeperTestSuite) TestMsgSetLossesInvalidUnauthorized() {
+	ctx, msgServer := s.ctx, s.msgServer
+	require := s.Require()
+
+	// Mock setup for addresses
+	reputerAddr := nonAdminAccounts[0].String()
+	workerAddr := sdk.AccAddress(PKS[1].Address()).String()
+
+	// Create a MsgSetLosses message
+	weightMsg := &state.MsgSetLosses{
+		Sender: reputerAddr,
+		Weights: []*state.Weight{
+			{
+				TopicId: 1,
+				Reputer: reputerAddr,
+				Worker:  workerAddr,
+				Weight:  cosmosMath.NewUint(100),
+			},
+		},
+	}
+
+	_, err := msgServer.SetWeights(ctx, weightMsg)
+	require.ErrorIs(err, state.ErrNotInReputerWhitelist, "SetWeights should return an error")
+}
