+++ x/emissions/keeper/msgserver/msg_server_params.go
@@ -1,0 +1,72 @@
+package msgserver
+
+import (
+	"context"
+
+	state "github.com/allora-network/allora-chain/x/emissions"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+func (ms msgServer) UpdateParams(ctx context.Context, msg *state.MsgUpdateParams) (*state.MsgUpdateParamsResponse, error) {
+	sender, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return nil, err
+	}
+	isAdmin, err := ms.k.IsWhitelistAdmin(ctx, sender)
+	if err != nil {
+		return nil, err
+	}
+	if !isAdmin {
+		return nil, state.ErrNotWhitelistAdmin
+	}
+	existingParams, err := ms.k.GetParams(ctx)
+	if err != nil {
+		return nil, err
+	}
+	// every option is a repeated field, so we interpret an empty array as "make no change"
+	newParams := msg.Params
+	if len(newParams.Version) == 1 {
+		existingParams.Version = newParams.Version[0]
+	}
+	if len(newParams.EpochLength) == 1 {
+		existingParams.EpochLength = newParams.EpochLength[0]
+	}
+	if len(newParams.EmissionsPerEpoch) == 1 {
+		existingParams.EmissionsPerEpoch = newParams.EmissionsPerEpoch[0]
+	}
+	if len(newParams.MinTopicUnmetDemand) == 1 {
+		existingParams.MinTopicUnmetDemand = newParams.MinTopicUnmetDemand[0]
+	}
+	if len(newParams.MaxTopicsPerBlock) == 1 {
+		existingParams.MaxTopicsPerBlock = newParams.MaxTopicsPerBlock[0]
+	}
+	if len(newParams.MinRequestUnmetDemand) == 1 {
+		existingParams.MinRequestUnmetDemand = newParams.MinRequestUnmetDemand[0]
+	}
+	if len(newParams.MaxMissingInferencePercent) == 1 {
+		existingParams.MaxMissingInferencePercent = newParams.MaxMissingInferencePercent[0]
+	}
+	if len(newParams.RequiredMinimumStake) == 1 {
+		existingParams.RequiredMinimumStake = newParams.RequiredMinimumStake[0]
+	}
+	if len(newParams.RemoveStakeDelayWindow) == 1 {
+		existingParams.RemoveStakeDelayWindow = newParams.RemoveStakeDelayWindow[0]
+	}
+	if len(newParams.MinRequestCadence) == 1 {
+		existingParams.MinRequestCadence = newParams.MinRequestCadence[0]
+	}
+	if len(newParams.MinLossCadence) == 1 {
+		existingParams.MinLossCadence = newParams.MinLossCadence[0]
+	}
+	if len(newParams.MaxInferenceRequestValidity) == 1 {
+		existingParams.MaxInferenceRequestValidity = newParams.MaxInferenceRequestValidity[0]
+	}
+	if len(newParams.MaxRequestCadence) == 1 {
+		existingParams.MaxRequestCadence = newParams.MaxRequestCadence[0]
+	}
+	err = ms.k.SetParams(ctx, existingParams)
+	if err != nil {
+		return nil, err
+	}
+	return &state.MsgUpdateParamsResponse{}, nil
+}
