+++ x/emissions/keeper/msgserver/msg_server_requests.go
@@ -1,0 +1,114 @@
+package msgserver
+
+import (
+	"context"
+
+	cosmosMath "cosmossdk.io/math"
+	"github.com/allora-network/allora-chain/app/params"
+	state "github.com/allora-network/allora-chain/x/emissions"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+///
+/// REQUESTS
+///
+
+func (ms msgServer) RequestInference(ctx context.Context, msg *state.MsgRequestInference) (*state.MsgRequestInferenceResponse, error) {
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+
+	for _, requestItem := range msg.Requests {
+		request := state.CreateNewInferenceRequestFromListItem(msg.Sender, requestItem)
+		// 1. check the topic is valid
+		topicExists, err := ms.k.TopicExists(ctx, request.TopicId)
+		if err != nil {
+			return nil, err
+		}
+		if !topicExists {
+			return nil, state.ErrInvalidTopicId
+		}
+		requestId, err := request.GetRequestId()
+		if err != nil {
+			return nil, err
+		}
+		// 2. check the request isn't already in the mempool
+		requestExists, err := ms.k.IsRequestInMempool(ctx, request.TopicId, requestId)
+		if err != nil {
+			return nil, err
+		}
+		if requestExists {
+			return nil, state.ErrInferenceRequestAlreadyInMempool
+		}
+		// 3. Check the BidAmount is greater than the price per request
+		if request.BidAmount.LT(request.MaxPricePerInference) {
+			return nil, state.ErrInferenceRequestBidAmountLessThanPrice
+		}
+		// 4. Check the timestamp valid until is in the future
+		timeNow := uint64(sdkCtx.BlockTime().Unix())
+		if request.TimestampValidUntil < timeNow {
+			return nil, state.ErrInferenceRequestTimestampValidUntilInPast
+		}
+		// 5. Check the timestamp validity is no more than the maximum allowed time in the future
+		maxInferenceRequestValidity, err := ms.k.GetParamsMaxInferenceRequestValidity(ctx)
+		if err != nil {
+			return nil, err
+		}
+		if request.TimestampValidUntil > timeNow+maxInferenceRequestValidity {
+			return nil, state.ErrInferenceRequestTimestampValidUntilTooFarInFuture
+		}
+		if request.Cadence != 0 {
+			// 6. Check the cadence is either 0, or greater than the minimum fastest cadence allowed
+			minFastestAllowedCadence, err := ms.k.GetParamsMinRequestCadence(ctx)
+			if err != nil {
+				return nil, err
+			}
+			if request.Cadence < minFastestAllowedCadence {
+				return nil, state.ErrInferenceRequestCadenceTooFast
+			}
+			// 7. Check the cadence is no more than the maximum allowed slowest cadence
+			maxSlowestAllowedCadence, err := ms.k.GetParamsMaxRequestCadence(ctx)
+			if err != nil {
+				return nil, err
+			}
+			if request.Cadence > maxSlowestAllowedCadence {
+				return nil, state.ErrInferenceRequestCadenceTooSlow
+			}
+		}
+		// 8. Check the cadence is not greater than the timestamp valid until
+		if timeNow+request.Cadence > request.TimestampValidUntil {
+			return nil, state.ErrInferenceRequestWillNeverBeScheduled
+		}
+		MinRequestUnmetDemand, err := ms.k.GetParamsMinRequestUnmetDemand(ctx)
+		if err != nil {
+			return nil, err
+		}
+		// Check that the request isn't spam by checking that the amount of funds it bids is greater than a global minimum demand per request
+		if request.BidAmount.LT(MinRequestUnmetDemand) {
+			return nil, state.ErrInferenceRequestBidAmountTooLow
+		}
+		// 9. Check sender has funds to pay for the inference request
+		// bank module does this for us in module SendCoins / subUnlockedCoins so we don't need to check
+		// 10. Send funds
+		senderAddr, err := sdk.AccAddressFromBech32(request.Sender)
+		if err != nil {
+			return nil, err
+		}
+		amountInt := cosmosMath.NewIntFromBigInt(request.BidAmount.BigInt())
+		coins := sdk.NewCoins(sdk.NewCoin(params.DefaultBondDenom, amountInt))
+		err = ms.k.SendCoinsFromAccountToModule(ctx, senderAddr, state.AlloraRequestsModuleName, coins)
+		if err != nil {
+			return nil, err
+		}
+		// 11. record the number of tokens sent to the module account
+		err = ms.k.SetRequestDemand(ctx, requestId, request.BidAmount)
+		if err != nil {
+			return nil, err
+		}
+		// 12. Write request state into the mempool state
+		request.LastChecked = timeNow
+		err = ms.k.AddToMempool(ctx, *request)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return &state.MsgRequestInferenceResponse{}, nil
+}
